move renderConversationForModel to a dedicated file to break a dependency cycle
